+++ rosu-v2/src/model/score_.rs
@@ -1,8 +1,8 @@
 use super::{
-    beatmap::{Beatmap, BeatmapsetCompact},
+    beatmap::{BeatmapExtended, Beatmapset},
     mods::GameMods,
     serde_,
-    user_::UserCompact,
+    user_::User,
     GameMode, Grade,
 };
 use crate::{mods, prelude::ModeAsSeed, request::GetUser, Osu};
@@ -34,7 +34,7 @@ pub struct BeatmapUserScore {
 }
 
 impl BeatmapUserScore {
-    /// Request the [`User`](crate::model::user::User) of the score
+    /// Request the [`UserExtended`](crate::model::user::UserExtended) of the score
     #[inline]
     pub fn get_user<'o>(&self, osu: &'o Osu) -> GetUser<'o> {
         self.score.get_user(osu)
@@ -60,12 +60,12 @@ pub struct Score {
         feature = "serialize",
         serde(rename = "beatmap", skip_serializing_if = "Option::is_none")
     )]
-    pub map: Option<Beatmap>,
+    pub map: Option<BeatmapExtended>,
     #[cfg_attr(
         feature = "serialize",
         serde(rename = "beatmapset", skip_serializing_if = "Option::is_none")
     )]
-    pub mapset: Option<BeatmapsetCompact>,
+    pub mapset: Option<Beatmapset>,
     pub mode: GameMode,
     pub mods: GameMods,
     pub perfect: bool,
@@ -81,7 +81,7 @@ pub struct Score {
     pub score_id: Option<u64>,
     pub statistics: ScoreStatistics,
     #[cfg_attr(feature = "serialize", serde(skip_serializing_if = "Option::is_none"))]
-    pub user: Option<UserCompact>,
+    pub user: Option<User>,
     pub user_id: u32,
     #[cfg_attr(feature = "serialize", serde(skip_serializing_if = "Option::is_none"))]
     pub weight: Option<ScoreWeight>,
@@ -102,9 +102,9 @@ impl<'de> Deserialize<'de> for Score {
             map_id: u32,
             max_combo: u32,
             #[serde(rename = "beatmap")]
-            map: Option<Beatmap>,
+            map: Option<BeatmapExtended>,
             #[serde(rename = "beatmapset")]
-            mapset: Option<BeatmapsetCompact>,
+            mapset: Option<Beatmapset>,
             #[serde(alias = "ruleset_id")]
             mode: GameMode,
             mods: Box<RawValue>,
@@ -119,7 +119,7 @@ impl<'de> Deserialize<'de> for Score {
             #[serde(rename = "best_id")]
             score_id: Option<u64>,
             statistics: ScoreStatistics,
-            user: Option<UserCompact>,
+            user: Option<User>,
             user_id: u32,
             weight: Option<ScoreWeight>,
         }